Add special rule to make unit tests easier
Since "assume" commands only take a single term, in order to have
premise clauses in tests, we were using an "assume" command with an "or"
term followed by a step that uses the "or" rule to unwrap it into to a
clause. This commit introduces the "trust_me" rule that simply returns
true for every clause. Of course, this rule can only be used in tests.