Optimize `TermPool::apply_substitutions`
Previously, this method was traversing terms as a tree in some cases. If
the term does not contain any variable to be substituted, the result of
calling `apply_substitutions` on it is the term itself. To avoid
polluting the given substitutions hash map, this method would skip
adding these symmetric substitutions to it (this was added in 6f39654).
However, since this hash map was also being used as the cache for the
traversal, terms like this would be traversed without a cache, which is
very inefficient. Now, this method creates an empty hash map to serve as
cache, and does not mutate the given substitutions hash map. This has
the downside that previous calls to `apply_substitutions` will no longer
be cached for when a new call with the same term happens, but overall
performance improves substantially with this change.