Change how integer literals are parsed depending on logic
This commit makes it so the parser will read "set-logic" commands from
the original problem file, and, if the logic contains real numbers but
not integers, interpret all integer literals as reals. That is, "1"
would be interpreted as "1.0". This is enough to correctly parse terms
in "assert" commands, and fixes the sort error problems we were having.
This reverts many of the changes introduced in 6c55ecc.