Fix bug in alpha equivalence checking
We were not popping the frame off of the alpha equivalence checker when
encountering binders that were trivially unequal. This could leave the
checker in a corrupted state. In most cases, returning `false` for a
subterm would mean that the terms as a whole were false as well, so this
state wouldn't affect the results. However, when checking an equality
that could be flipped, returning `false` may not end the execution of
the alpha equivalence algorithm as a whole, and the corrupted state
could then lead to incorrect results.